Improvements to day 17
A post on the subreddit explained that zero is always at the end of the buffer.
Proof by induction.
At the start, zero is at the end of the buffer. (base case).
For some loop n, assume that zero is at the end of the buffer, index i. The spinlock must land on p <= i. inserting at p <= will keep zero at the end of the buffer. (inductive case)
Given this, we can substitute 1 for wrap(zero_pos, 1, buffer_len) ,
substitute buffer_len for zero_pos, and simplify.